page ファイルの削除
