--- packages/frontend/src/pages/favorite.tsx
@@ -1,129 +1,0 @@
-import { Center, Group, Stack } from "@mantine/core";
-import { useRouter, useSearch } from "@tanstack/react-router";
-import { useState } from "react";
-import FilteringModalButton from "@/features/filtering/components/FilteringModalButton";
-import QuizControllBar from "@/features/quiz/components/QuizControllBar";
-import QuizHiddenAnswerButton from "@/features/quiz/components/QuizHiddenAnswerButton";
-import QuizList from "@/features/quiz/components/QuizList";
-import QuizPagination from "@/features/quiz/components/QuizPagination";
-import QuizShuffleButton from "@/features/quiz/components/QuizShuffleButton";
-import QuizTransfarButton from "@/features/quiz/components/QuizTransfarButton";
-import { $api } from "@/utils/client";
-
-export default function Favorite() {
-  const router = useRouter();
-  const search = useSearch({ from: "/favorite" });
-  const [activePage, setPage] = useState(1);
-  const [isHidden, setIsHidden] = useState(false);
-  const { data: quizzes } = $api.useQuery("get", "/quizzes", {
-    params: { query: search },
-  });
-  const { data: quizzesSize } = $api.useQuery("get", "/quizzes/size", {
-    params: { query: search },
-  });
-  const size =
-    !!quizzes && !!quizzes.length && !!quizzesSize ? quizzesSize.size : 0;
-
-  const toFilter = (
-    workbooks?: string | string[],
-    keyword?: string,
-    keywordOption?: number,
-    categories?: number | number[],
-    tags?: string | string[],
-    tagMatchAll?: boolean,
-    perPage?: number,
-  ) => {
-    setPage(1);
-    router.navigate({
-      to: "/favorite",
-      search: (old) => ({
-        ...old,
-        page: 1,
-        seed: undefined,
-        maxView: perPage,
-        wids: workbooks,
-        keyword,
-        keywordOption,
-        categories,
-        tags,
-        tagMatchAll,
-      }),
-      replace: true,
-    });
-  };
-
-  const toShuffle = (seed: number) => {
-    setPage(1);
-    router.navigate({
-      to: "/favorite",
-      search: (old) => ({
-        ...old,
-        page: 1,
-        seed,
-      }),
-      replace: true,
-    });
-  };
-
-  const changePage = (page: number) => {
-    setPage(page);
-    router.navigate({
-      to: "/favorite",
-      search: (old) => ({
-        ...old,
-        page,
-      }),
-      replace: true,
-    });
-  };
-
-  const toTransfar = () => {
-    router.navigate({
-      to: "/practice",
-      search: {
-        ...search,
-        isTransfer: true,
-      },
-    });
-  };
-
-  return (
-    <>
-      <QuizControllBar
-        p="sm"
-        total={size}
-        buttons={
-          <Group>
-            <FilteringModalButton onSubmit={toFilter} />
-            <QuizShuffleButton apply={toShuffle} />
-            <QuizHiddenAnswerButton
-              isHidden={isHidden}
-              onToggle={setIsHidden}
-            />
-            <QuizTransfarButton
-              apply={toTransfar}
-              disabled={quizzes?.length === 0}
-            />
-          </Group>
-        }
-        pagination={
-          <Stack gap={2}>
-            <Center mt="sm">
-              <QuizPagination
-                page={activePage}
-                total={!!search?.maxView ? Math.ceil(size / search.maxView) : 0}
-                setPage={changePage}
-              />
-            </Center>
-          </Stack>
-        }
-      />
-      <QuizList
-        quizzes={quizzes}
-        page={activePage}
-        perPage={search.maxView || 0}
-        isHidden={isHidden}
-      />
-    </>
-  );
-}
